+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Puzzle Shell Changelog
 
+## 1.0.0-alpha.6
+
+- Add rudimentary `pzsh-nav`, `pzsh-subnav` and `pzsh-nav-item` components for application navigation
+- Implement responsive `pzsh-banner` component
+
 ## 1.0.0-alpha.5
 
 - Add responsive `pzsh-menu-dropdown` component